+++ src/components/common/Context.jsx
@@ -1,0 +1,4 @@
+import React, { createContext } from 'react';
+
+
+export const Context = React.createContext({});
