--- src/components/common/TestContext.jsx
@@ -1,4 +1,0 @@
-import React, { createContext } from 'react';
-
-
-export const TestContext = React.createContext({});
