+++ src/components/test/Content.jsx
@@ -1,8 +1,7 @@
 import React, { useState, useEffect, useRef, useCallback, useMemo, useContext } from 'react';
 import styled from 'styled-components';
-import {TestContext} from '../common/TestContext';
+import {Context} from '../common/Context';
 import media from '../../lib/styles/media';
-
 const slideWidth = 400;
 
 const SlideContent = styled.div`
@@ -32,6 +31,7 @@ const SlideText = styled.p`
   display: table-cell;
   text-align: center;
   vertical-align: middle;
+  font-family: 'jua';
   font-size: 20px;
   margin: 40px 0;
 `;
@@ -51,6 +51,7 @@ const SlideBtnButton = styled.button`
   margin: 20px auto;
   width: 100%;
   height: 65px;
+  font-family: 'jua';
   font-size: 16px;
   background: none;
   border: 1px solid #FF9C9C;
@@ -69,22 +70,13 @@ const Content = ({nextSlide, data}) => {
   // const [count, setCount] = useState(0);
   // count로 마지막 질문지 체크한 후, <Link> 컴포넌트 출력하는 함수 만들어서 넣기..
 
-  const {state, dispatch } = useContext(TestContext);
+  const {state, dispatch } = useContext(Context);
 
-  // const getAnswer = useCallback((id, isSelected) =>{
-  //   if(isSelected === "A"){
-  //     dispatch(id);
-  //   }else if(isSelected ==="B"){
-  //     dispatch(id);
-  //   }else{
-  //     console.log('isSelected is wrong');
-  //   }
-  // });
   
   const clickSelection = useCallback((id, isSelected)=>{
     nextSlide();
+    dispatch(id);
     console.log(state);
-    // getAnswer(id, isSelected);
   });
 
   return (
@@ -95,10 +87,10 @@ const Content = ({nextSlide, data}) => {
         </SlideText>
       </SlideTextWrap>
       <SlideBtnBox>
-        <SlideBtnButton type="button" onClick={()=>{clickSelection(data.id,"A"); dispatch({type:data.id}); }}>
+        <SlideBtnButton type="button" onClick={()=>{clickSelection({type:data.id},"A"); }}>
           {data.A}
         </SlideBtnButton>
-        <SlideBtnButton type="button" onClick={()=>{clickSelection(data.id,"B"); /*dispatch({type:data.id})*/}}>
+        <SlideBtnButton type="button" onClick={()=>{clickSelection({type:data.id},"B"); /*dispatch({type:data.id})*/}}>
           {data.B}
         </SlideBtnButton>
       </SlideBtnBox>
