+++ src/lib/styles/GlobalStyle.js
@@ -3,6 +3,10 @@ import reset from 'styled-reset';
 
 const GlobalStyle = createGlobalStyle`
   ${reset}
+  @font-face {
+    font-family: 'jua';
+    src: url(../../../public/font/BMJUA.ttf);
+  };
 `;
 
 export default GlobalStyle;
